fix: rename ng to angular
