+++ angular.json
@@ -3,21 +3,21 @@
   "version": 1,
   "newProjectRoot": "projects",
   "projects": {
-    "ng-ipfs-service": {
+    "angular-ipfs-service": {
       "projectType": "library",
-      "root": "projects/ng-ipfs-service",
-      "sourceRoot": "projects/ng-ipfs-service/src",
+      "root": "projects/angular-ipfs-service",
+      "sourceRoot": "projects/angular-ipfs-service/src",
       "prefix": "lib",
       "architect": {
         "build": {
           "builder": "@angular-devkit/build-angular:ng-packagr",
           "options": {
-            "tsConfig": "projects/ng-ipfs-service/tsconfig.lib.json",
-            "project": "projects/ng-ipfs-service/ng-package.json"
+            "tsConfig": "projects/angular-ipfs-service/tsconfig.lib.json",
+            "project": "projects/angular-ipfs-service/ng-package.json"
           },
           "configurations": {
             "production": {
-              "tsConfig": "projects/ng-ipfs-service/tsconfig.lib.prod.json"
+              "tsConfig": "projects/angular-ipfs-service/tsconfig.lib.prod.json"
             }
           },
           "allowedCommonJsDependencies": ["ipfs-core"]
@@ -25,22 +25,22 @@
         "test": {
           "builder": "@angular-devkit/build-angular:karma",
           "options": {
-            "main": "projects/ng-ipfs-service/src/test.ts",
-            "tsConfig": "projects/ng-ipfs-service/tsconfig.spec.json",
-            "karmaConfig": "projects/ng-ipfs-service/karma.conf.js"
+            "main": "projects/angular-ipfs-service/src/test.ts",
+            "tsConfig": "projects/angular-ipfs-service/tsconfig.spec.json",
+            "karmaConfig": "projects/angular-ipfs-service/karma.conf.js"
           }
         },
         "lint": {
           "builder": "@angular-eslint/builder:lint",
           "options": {
             "lintFilePatterns": [
-              "projects/ng-ipfs-service/**/*.ts",
-              "projects/ng-ipfs-service/**/*.html"
+              "projects/angular-ipfs-service/**/*.ts",
+              "projects/angular-ipfs-service/**/*.html"
             ]
           }
         }
       }
     }
   },
-  "defaultProject": "ng-ipfs-service"
+  "defaultProject": "angular-ipfs-service"
 }
