+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ng-ipfs-service",
+  "name": "angular-ipfs-service",
   "version": "1.1.5",
   "engines": {
     "node": ">=12",
