+++ projects/angular-ipfs-service/karma.conf.js
@@ -25,7 +25,10 @@ module.exports = function (config) {
       suppressAll: true, // removes the duplicated traces
     },
     coverageReporter: {
-      dir: require("path").join(__dirname, "../../coverage/ng-ipfs-service"),
+      dir: require("path").join(
+        __dirname,
+        "../../coverage/angular-ipfs-service"
+      ),
       subdir: ".",
       reporters: [
         { type: "html" },
