+++ projects/angular-ipfs-service/ng-package.json
@@ -1,6 +1,6 @@
 {
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/ng-ipfs-service",
+  "dest": "../../dist/angular-ipfs-service",
   "lib": {
     "entryFile": "src/public-api.ts"
   },
