+++ projects/angular-ipfs-service/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "ng-ipfs-service",
+  "name": "angular-ipfs-service",
   "version": "1.1.5",
   "peerDependencies": {
     "@angular/common": "^11.1.1",
@@ -15,12 +15,12 @@
     "IPFS",
     "Angular"
   ],
-  "homepage": "https://github.com/ocknamo/ng-ipfs-service",
+  "homepage": "https://github.com/ocknamo/angular-ipfs-service",
   "repository": {
     "type": "git",
-    "url": "https://github.com/ocknamo/ng-ipfs-service.git"
+    "url": "https://github.com/ocknamo/angular-ipfs-service.git"
   },
   "bugs": {
-    "url": "https://github.com/ocknamo/ng-ipfs-service/issues"
+    "url": "https://github.com/ocknamo/angular-ipfs-service/issues"
   }
 }
