+++ projects/angular-ipfs-service/src/lib/angular-ipfs.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { NgIpfsService } from './ng-ipfs.service';
+import { NgIpfsService } from './angular-ipfs.service';
 
 @NgModule({
   providers: [NgIpfsService],
