+++ projects/angular-ipfs-service/src/lib/angular-ipfs.service.ts
