+++ projects/angular-ipfs-service/src/public-api.ts
@@ -1,0 +1,6 @@
+/*
+ * Public API Surface of angular-ipfs-service
+ */
+
+export * from './lib/angular-ipfs.service';
+export * from './lib/angular-ipfs.module';
