+++ projects/angular-ipfs-service/tsconfig.lib.json
