+++ projects/angular-ipfs-service/tsconfig.lib.prod.json
