+++ projects/angular-ipfs-service/tsconfig.spec.json
