--- projects/ng-ipfs-service/src/public-api.ts
@@ -1,6 +1,0 @@
-/*
- * Public API Surface of ng-ipfs-service
- */
-
-export * from './lib/ng-ipfs.service';
-export * from './lib/ng-ipfs.module';
